+++ Server/auth/jwtutils.go
@@ -1,4 +1,4 @@
-package utils
+package auth
 
 import (
 	"time"
